build: add accounts app and database credentials
The `accounts` django app is created to store the details of three different
users: a household, an individual household member or an admin. A custom
user model, named as CustomAccount, is created for this purpose and
given to django use as such by populating django's AUTH_USER_MODEL
variable.

Additional environment variables for production are provided for
accessing the project's mysql database via a specific user's database
admin credentials.