add: a model and its manager for account creation
The `CustomAccount` django model is defined to represent the
`customaccounts` table in the project's database. It uses a manager to
interact with the database: make queries and receive responses.
The `CustomAccountManager` manager inherits from BaseUserManager and thus
has access to many useful methods for interacting with the database.
It uses its own custom methods for normal- and super- account creation,
mainly to use an email as a unique indentifier for users instead of a
username.